CPR-141 move validation into PNCIdentifier
