+++ src/test/kotlin/uk/gov/justice/digital/hmpps/personrecord/validate/PNCIdValidatorTest.kt
@@ -1,30 +1,16 @@
 package uk.gov.justice.digital.hmpps.personrecord.validate
 
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.extension.ExtendWith
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
-import org.mockito.Mock
-import org.mockito.junit.jupiter.MockitoExtension
-import uk.gov.justice.digital.hmpps.personrecord.service.TelemetryService
 import java.nio.file.Files
 import java.nio.file.Paths
 import kotlin.text.Charsets.UTF_8
 
-@ExtendWith(MockitoExtension::class)
 class PNCIdValidatorTest {
 
-  @Mock
-  lateinit var telemetryService: TelemetryService
-
-  private lateinit var pncIdValidator: PNCIdValidator
-
-  @BeforeEach
-  fun setUp() {
-    pncIdValidator = PNCIdValidator(telemetryService)
-  }
+  private var pncIdValidator: PNCIdValidator = PNCIdValidator()
 
   @ParameterizedTest
   @ValueSource(
